+++ packages/utils/src/index.ts
@@ -2,6 +2,7 @@ export { once, untilAborted } from "./abortable";
 export * from "./async";
 export * from "./binary";
 export * from "./color";
+export * from "./crash-report";
 export * from "./dirs";
 export * from "./env";
 export * from "./fetch-retry";